@@ -1,17 +1,14 @@
 package com.tristanruecker.interviewexampleproject.models.objects.entities;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.persistence.Column;
+import jakarta.persistence.MappedSuperclass;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import org.springframework.lang.NonNull;
 
-import javax.persistence.Column;
-import javax.persistence.MappedSuperclass;
-
 @Data
 @NoArgsConstructor
-@EqualsAndHashCode(callSuper = false)
 @MappedSuperclass
 public class UserEmailAndPassword extends AbstractDatabaseObject {
 
